Update private key files in deployment scripts.
These are now stored in a common location ~/.ssh/opentree. In addition,
the GITHUB_CLIENT_SECRET file will vary with the final public-facing
domain, so we now expect to find a file with this domain in its
filename, for example:

~/.ssh/opentree/GITHUB_CLIENT_SECRET-dev.opentreeoflife.org

This domain name will be stripped when the file is pushed to our webapp
server, so no other changes are required.