Support for a 4th arg to git_refresh for dir
The multi-repo version of api needs a
parent directory that will be the parent
of each "shard" of the phylesystem. So
I added an optional 4th arg to git_refresh so
that the api could checkout the phylesystem
repos to repo/phylesystem_par/phyelsystem
rather than repo/phylesystem
The install-api.sh depends on the local-dep
branch of api. So it should not be merged to
master until that merge happens.