2 accessor for github_oauth and bug fix
Previously run_tests.sh would override the command line
(and test.conf) value for host.allowwrite with "false".

The 2 new accessors are just sugar for making sure
that you are not running write-only tests in readonly
mode, or against the production api server, and that you
have the (necessary) env variable.