Moving all daemon management to a separate script.
This will run as admin for the required privileges, and we'll choose
which daemons to add/update based on the components being deployed (or
the command being run) in the main push script.

We're starting out by testing the installation of a web2py session
sweeper for the phylesystem-api component.