Manage neo4j instances with systemd
This replaces most of the ad-hoc starting/stopping of neo4j instances
with more robust (re)loading and (re)starting of persistent daemons, so
that these services will be restored after reboots. Systemd will also
attempt to relaunch any failed neo4j instance.

NOTE that there might a one-time error when deploying these changes,
which will be reported during deployment as "control process exited with
error code". This happens when the server is still running an older
(explicitly started) neo4j instance on the intended port.

In this case, you'll need to
 - find its PID using `ps -axwww | grep neo4j | grep -v grep`,
 - kill the offending process with `kill -9 $PID`, then
 - retry the deploying this neo4j instance.