Automate changing canonical hostname
Apparently this was formerly a manual step? Now we make it official,
setting the `hostname` and `/etc/hostname` to survive reboots of the
deployed server. This is necessary for apache to trust some
configuration.

Any chance this goes wrong if we have deployed multiple systems to
the same server? e.g. `devtree.opentreeoflife.org` and
`devapi.opentreeoflife.org`. Comments welcome!